Improved? Models
Reworked the way an Exercise is built. It now has (read: is supposed to
have) the possibility to have multiple Questions (so that different
types of exercises can be created), each of which poses a question and
can have multiple Answers. An Answer consists of an answer to a single
question, the possibility of having multiple answers allows for
different 'correct' replies.