Inheritance issues
Attempted to solve the issue of Sequence subclasses being unable to
define 'scene' in their own way to let them establish a ForeignKey
relationship with a Scene. I failed. Badly.

Even when removing the inheritance, Putty still said the local field
'scene' of a a subclass clashed with the field of the base class
Sequence. Weird as heck, as they weren't base classes anymore.