UserIntel; tracks user performance
Added the model UserIntel, and accompanying other things for in the
views and serializers.
Has ForeignKey with Scene to let a Scene have multiple instances of this
model. It has a maximum intel value that can be scored for an event, and
has a variable that holds the earned value of the user for that event.

Also holds an integer corresponding to the number of the event it
relates to, and the current user.