Update server-side code (#32)
* UserIntel; tracks user performance

Added the model UserIntel, and accompanying other things for in the
views and serializers.
Has ForeignKey with Scene to let a Scene have multiple instances of this
model. It has a maximum intel value that can be scored for an event, and
has a variable that holds the earned value of the user for that event.

Also holds an integer corresponding to the number of the event it
relates to, and the current user.

* Add image to stories

* Added image to Scene

* Added image to Scene

* Add migrations

* Changed UserIntel into Score

Also removed unnecessary fields that are supposed to be in Scene

* Migration?

This suddenly appeared.

* Started on Story functionality

Added the scoreMax and scoreReq to Scene.
Also added GET functionality to Scene to retrieve these two. Not sure if
this works correctly.

Also started on the function for Story that is supposed to return the
maximum achievable score. Kinda failing at the moment due to for-loop
shenenigans... <_<

* Commented the CalcTotalScore function

This way the code is usable. Pretty important stuff.

* More migrate shenenigans

* Added Serializer fields of scoreMax&Req to Scene

More fiddling around, can't get the for loop to work...

* calcTotalScoreMax function

Think I got the function down now. It's called calcTotalScoreMax and
it's -supposed- to return the maximum achievable score, as determined by
the scoreMax variables in Scene, which has a ForeignKey to Story (in
which this function is located)

* User struggles

Attempted to get the link to a user working; failed.
Efforts made are in the getUser function in the serializers folder.

* Install Django Extensions (for easy db reset)

* Remove redundant code and reset migrations

* Add code style check for python (server)

* Add verbal score string

* Implement full scores serializer

* Implement scores API